Split and complete UI plan test coverage suites
